refactor: merge asset_viz submodule into dartwork_mpl.diagnostics (#57)
Closes the final v0.2.0 folder-restructuring follow-up tracked as
issue #57. The four asset-inspection helpers (`classify_colormap`,
`plot_colormaps`, `plot_colors`, `plot_fonts`) now live in the new
top-level `dartwork_mpl.diagnostics` module; the legacy
`dartwork_mpl.asset_viz` subpackage is a thin deprecation shim that
re-exports the same four names and emits a `DeprecationWarning` on
import, mirroring the `helpers.formatting → helpers.labels` pattern
from #47.

Changes:
- New `src/dartwork_mpl/diagnostics.py` bundling the colormap /
  color-library / font helpers previously spread across
  `asset_viz/_cmap.py`, `_color.py`, `_font.py`. Intra-package
  import (`from ..cmap import ensure_loaded` → `from .cmap import
  ensure_loaded`) and the font/color asset paths
  (`Path(__file__).parent.parent` → `Path(__file__).parent`) were
  adjusted for the new module depth.
- `asset_viz/__init__.py` shrunk to a shim that re-exports from
  `diagnostics` and warns; `_cmap.py`, `_color.py`, `_font.py`
  removed.
- `dartwork_mpl/__init__.py` and `dartwork_mpl/explore.py` both
  import the four helpers from `.diagnostics` (canonical) instead of
  `.asset_viz`. Public API contract (`dm.classify_colormap` etc.) is
  unchanged.
- `tests/test_diagnostics.py` (new, 14 tests) exercises the
  canonical path plus a `TestTopLevelReexport` class pinning
  `dm.<name>` and `dm.explore.<name>` identity with
  `dm.diagnostics.<name>`.
- `tests/test_asset_viz.py` rewritten as a deprecation regression
  (3 tests): legacy import still works, emits
  `DeprecationWarning`, and re-exports the same callables.
- CHANGELOG Added / Changed / Deprecated entries; `docs/development/
  folder-restructuring.md` "Outstanding follow-up" updated to mark
  #57 shipped.

Test: 578 passed / 3 skipped (was 575/3).
Lint/format/mypy: all green.
