+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - **Dedicated tests for four previously-uncovered public modules**: `tests/test_figure.py` (9 tests covering `dm.subplots` and `dm.figure`, including style/figsize/dpi forwarding and invalid-style handling), `tests/test_formatting.py` (15 tests, one per axis formatter plus parametrised decimals and rotation), `tests/test_spines.py` (9 tests, one per spine/grid helper with concrete spine-visibility assertions), and `tests/test_validate_enhanced.py` (7 tests covering `validate_with_fixes`, `get_fix_suggestions`, `check_agent_requirements`, `generate_validation_report`, and the new `dm.validate_with_fixes` top-level alias). Suite goes from 534 to 574 passing tests.
 - **Gallery: six new single-plot spine / grid examples** under `docs/examples_source/01_styling_and_themes/` — `plot_spine_minimal.py`, `plot_spine_visibility.py`, `plot_spine_styling.py`, `plot_grid_customization.py`, `plot_spine_publication_styles.py`, `plot_spine_dark_theme.py`, `plot_spine_dashboard.py`. Each renders exactly one figure, per the repo's one-plot-per-file convention.
 - **Gallery: six new single-plot helpers examples** under `docs/examples_source/05_advanced_components/` — `plot_helpers_data_validation.py`, `plot_helpers_color_selection.py`, `plot_helpers_labels.py` (renamed from `plot_helpers_formatting.py` for consistency with the new `dm.helpers.labels` submodule), `plot_helpers_quality.py`, `plot_helpers_io.py`, `plot_helpers_workflow.py`. Each renders exactly one figure and covers one `dm.helpers` submodule (plus an end-to-end workflow demo), replacing the monolithic six-figure `plot_helpers_usage.py`.
+- **`dartwork_mpl.diagnostics`** — new canonical home for the four asset-inspection helpers (`classify_colormap`, `plot_colormaps`, `plot_colors`, `plot_fonts`). Closes out the final v0.2.0 folder-restructuring follow-up tracked as issue #57. The functions themselves are unchanged; only their physical module path moved. All four names remain reachable via `dartwork_mpl.<name>` (top-level), `dartwork_mpl.explore.<name>` (re-export), and — with a `DeprecationWarning` — via the legacy `dartwork_mpl.asset_viz.<name>` path.
+- **`tests/test_diagnostics.py`** — 14 tests covering the canonical `dartwork_mpl.diagnostics` import path (functional behaviour of the four helpers plus a `TestTopLevelReexport` class pinning that `dm.<name>` and `dm.explore.<name>` resolve to the same objects as `dm.diagnostics.<name>`). Migrated and expanded from the legacy `tests/test_asset_viz.py`.
 
 ### Changed
 
@@ -27,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - **`docs/examples_source/` rewritten to one-plot-per-file convention** (first pass): earlier bundle files (`single_plot_example.py`, `single_plot_korean.py`, monolithic gallery entries) were split so each file renders exactly one `Figure`. Also de-domained example captions and data labels.
 - **Gallery: `plot_waterfall_bridge.py` generalised to an energy-balance walk** instead of Year-over-Year profit change, so the gallery entry leads with a neutral domain while the narrative text still mentions other valid domains (mass balance, population flow, etc.).
 - **Gallery: financial dashboard replaced by a set of single-chart sensor examples** (`docs/examples_source/06_real_world_dashboards/plot_sensor_*.py`), removing the last standalone finance-branded gallery entry and keeping the "real-world dashboard" theme with a neutral subject.
+- **Canonical import path for asset diagnostics is now `dartwork_mpl.diagnostics`.** The top-level `__init__.py` and `dartwork_mpl.explore` both re-export `classify_colormap`, `plot_colormaps`, `plot_colors`, and `plot_fonts` from the new `diagnostics` module (previously they came from `asset_viz`). Public API contract (`dm.classify_colormap(...)`, `dm.plot_colormaps(...)`, etc.) is unchanged — only the physical module path moved. `__init__.py`'s internal import block comment was updated to reflect the new source module.
 
 ### Fixed
 
@@ -41,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 - **`dartwork_mpl.helpers.formatting`** — deprecated alias for `dartwork_mpl.helpers.labels`. Importing from the old path still works but emits a `DeprecationWarning` pointing at the new submodule. Renamed to resolve the long-standing name clash with the top-level `dartwork_mpl.formatting` module (which houses the `format_axis_*` tick formatters). Plan to remove the deprecated shim in a future major release.
+- **`dartwork_mpl.asset_viz`** — deprecated alias for `dartwork_mpl.diagnostics`. The subpackage is now a thin shim: `asset_viz/__init__.py` re-exports the four helpers from `dartwork_mpl.diagnostics` and emits a `DeprecationWarning` on import. `asset_viz/_cmap.py`, `asset_viz/_color.py`, and `asset_viz/_font.py` were removed (their contents are now in `src/dartwork_mpl/diagnostics.py`). Plan to remove the deprecated shim in a future major release alongside `dm.agent_utils`, `dm.xplot`, and `dm.helpers.formatting`.
 
 ### Removed
